[Improve][Connector-V2][File] Move empty file list read strategy tests to AbstractReadStrategyTest
